Move valueSetsReachingValue into ValueSetSet class
It seems to make more sense here, considering what it does and what data
it works with.
