+++ Annotator.hh
@@ -7,7 +7,6 @@
 #include "FindStructElements.hh"
 #include "IIGlueReader.hh"
 #include "LengthInfo.hh"
-#include "ValueSetsReachingValue.hh"
 
 #include <llvm/IR/Module.h>
 #include <llvm/Pass.h>
