+++ FindLengthChecks.cc
@@ -5,7 +5,6 @@
 #include "FindLengthChecks.hh"
 #include "IIGlueReader.hh"
 #include "SRA/SymbolicRangeAnalysis.h"
-#include "ValueSetsReachingValue.hh"
 
 #include <boost/algorithm/cxx11/any_of.hpp>
 #include <boost/foreach.hpp>
